[MNG-8084] Change six remaining imports of `ModelVersionParser`. This commit aims to complete d075fe7e85d26d86ad25a963274cd9c5b910c56a, which moved `ModelVersionParser` interface from the `org.apache.maven.model.version` package to `org.apache.maven.api.services.model`. Completing this change is necessary for building `maven-plugin-tools` or `maven-plugin-testing`.
This commit deletes the old interface for making sure that there is no accidental use.
The old and new interfaces were identical except for the package names and the imports.
As a result of this change, the following implementation classes are identical except
for package names and imports:

* `maven-resolver-provider/src/main/java/org/apache/maven/repository/internal/DefaultModelVersionParser.java`
* `maven-api-impl/src/main/java/org/apache/maven/internal/impl/DefaultModelVersionParser.java`
