+++ maven-resolver-provider/src/test/java/org/apache/maven/repository/internal/ModelVersionParserTest.java
@@ -20,7 +20,7 @@ package org.apache.maven.repository.internal;
 
 import org.apache.maven.api.VersionConstraint;
 import org.apache.maven.api.services.VersionParserException;
-import org.apache.maven.model.version.ModelVersionParser;
+import org.apache.maven.api.services.model.ModelVersionParser;
 import org.eclipse.aether.util.version.GenericVersionScheme;
 import org.junit.jupiter.api.Test;
